Minor Improvements

+++ Software/RaspberryPi/sensors-actuators/src/main/java/com/subutai/nova/arduino/command/Commands.java
@@ -2,19 +2,19 @@ package com.subutai.nova.arduino.command;
 
 public class Commands {
 
-    public static final byte RES0 = 0;
-    public static final byte RES1 = 1;
-    public static final byte RES2 = 2;
-    public static final byte RES3 = 3;
-    public static final byte RES4 = 4;
-    public static final byte RES5 = 5;
-    public static final byte RES6 = 6;
-    public static final byte RES7 = 7;
-    public static final byte RES8 = 8;
-    public static final byte RES9 = 9;
+    public static final int RES0 = 0;
+    public static final int RES1 = 1;
+    public static final int RES2 = 2;
+    public static final int RES3 = 3;
+    public static final int RES4 = 4;
+    public static final int RES5 = 5;
+    public static final int RES6 = 6;
+    public static final int RES7 = 7;
+    public static final int RES8 = 8;
+    public static final int RES9 = 9;
 
-    public static final byte REQ_YPR = 10;
-    public static final byte STREAM_YPR = 11;
-    public static final byte LED_ON = 12;
-    public static final byte LED_OFF = 13;
+    public static final int REQ_YPR = 10;
+    public static final int STREAM_YPR = 11;
+    public static final int LED_ON = 12;
+    public static final int LED_OFF = 13;
 }
