+++ Software/RaspberryPi/sensors-actuators/src/main/java/com/subutai/nova/arduino/command/list/LedOn.java
@@ -10,7 +10,7 @@ public class LedOn extends ArduinoCommand {
     private int color;
 
     public LedOn(int color) {
-        super.id = Commands.LED_ON;
+        super.id = (byte) Commands.LED_ON;
         this.color = color;
     }
